Issue #9 , sending html directly to email
